requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
openai==1.12.0
supabase==2.3.4
pydantic==2.5.2
python-dotenv==1.0.0
httpx==0.25.2
